[FIX/#104] Main 패키지 수정
