--- app/src/main/java/org/sopt/santamanitto/main/MyManittoInfo.kt
@@ -1,6 +1,0 @@
-package org.sopt.santamanitto.main
-
-data class MyManittoInfo(
-        val manittoName: String,
-        val mission: String?
-)
