+++ app/src/main/java/org/sopt/santamanitto/main/list/AdapterModule.kt
@@ -1,4 +1,4 @@
-package org.sopt.santamanitto.main
+package org.sopt.santamanitto.main.list
 
 import dagger.Module
 import dagger.Provides
