+++ app/src/main/java/org/sopt/santamanitto/main/list/MyManittoModel.kt
@@ -1,0 +1,6 @@
+package org.sopt.santamanitto.main.list
+
+data class MyManittoModel(
+        val manittoName: String,
+        val mission: String?
+)
