+++ app/src/main/java/org/sopt/santamanitto/main/list/RemovedMyManttioViewHolder.kt
@@ -1,4 +1,4 @@
-package org.sopt.santamanitto.main.viewholder
+package org.sopt.santamanitto.main.list
 
 import android.view.ViewGroup
 import org.sopt.santamanitto.R
@@ -6,16 +6,13 @@ import org.sopt.santamanitto.databinding.ItemMymanittoRemovedBinding
 import org.sopt.santamanitto.room.data.MyManitto
 import org.sopt.santamanitto.view.recyclerview.BaseViewHolder
 
-class ExpiredMyManittoViewHolder(
+class RemovedMyManttioViewHolder(
     parent: ViewGroup,
-    enterListener: ((roomId: Int, isMatched: Boolean, isFinished: Boolean) -> Unit)?,
     removeListener: ((roomId: Int) -> Unit)?
 )
     : BaseViewHolder<MyManitto, ItemMymanittoRemovedBinding>(R.layout.item_mymanitto_removed, parent) {
 
     private val removeButton = binding.textviewMymanittoremovedButton
-    private val title = binding.textviewMymanittoremovedTitle
-    private val content = binding.textviewMymanittoremovedContent
 
     private var roomId = -1
 
@@ -25,16 +22,9 @@ class ExpiredMyManittoViewHolder(
                 it(roomId)
             }
         }
-        enterListener?.let {
-            binding.root.setOnClickListener {
-                it(roomId, false, false)
-            }
-        }
-        content.text = itemView.context.getString(R.string.mymanitto_epired_content)
     }
 
     override fun bind(data: MyManitto) {
-        title.text = data.roomName
         roomId = data.roomId
     }
 }
